[#77] refactoring. seprate files.
